+++ app/src/main/java/de/westnordost/streetcomplete/controls/MainMenuButtonFragment.kt
@@ -3,32 +3,30 @@ package de.westnordost.streetcomplete.controls
 import android.net.ConnectivityManager
 import android.os.Bundle
 import android.view.View
-import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
 import androidx.core.content.getSystemService
 import androidx.fragment.app.Fragment
 import de.westnordost.osmapi.map.data.BoundingBox
-import de.westnordost.streetcomplete.ApplicationConstants
 import de.westnordost.streetcomplete.Injector
 import de.westnordost.streetcomplete.R
 import de.westnordost.streetcomplete.data.download.DownloadController
 import de.westnordost.streetcomplete.ktx.toast
-import de.westnordost.streetcomplete.map.MainFragment
-import de.westnordost.streetcomplete.util.area
-import de.westnordost.streetcomplete.util.asBoundingBoxOfEnclosingTiles
-import de.westnordost.streetcomplete.util.enclosingBoundingBox
 import kotlinx.android.synthetic.main.fragment_main_menu_button.view.*
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import javax.inject.Inject
-import kotlin.math.PI
-import kotlin.math.sqrt
 
 /** Fragment that shows the main menu button and manages its logic */
 class MainMenuButtonFragment : Fragment(R.layout.fragment_main_menu_button),
     CoroutineScope by CoroutineScope(Dispatchers.Main) {
 
     @Inject internal lateinit var downloadController: DownloadController
+
+    interface Listener {
+        fun getDownloadArea(): BoundingBox?
+    }
+
+    private val listener: Listener? get() = parentFragment as? Listener ?: activity as? Listener
 
     /* --------------------------------------- Lifecycle ---------------------------------------- */
 
@@ -62,43 +60,20 @@ class MainMenuButtonFragment : Fragment(R.layout.fragment_main_menu_button),
     }
 
     private fun downloadDisplayedArea() {
-        val displayArea = (requireParentFragment() as MainFragment).mapFragment?.getDisplayedArea()
-        if (displayArea == null) {
-            context?.toast(R.string.cannot_find_bbox_or_reduce_tilt, Toast.LENGTH_LONG)
-        } else {
-            val enclosingBBox = displayArea.asBoundingBoxOfEnclosingTiles(ApplicationConstants.QUEST_TILE_ZOOM)
-            val areaInSqKm = enclosingBBox.area() / 1000000
-            if (areaInSqKm > ApplicationConstants.MAX_DOWNLOADABLE_AREA_IN_SQKM) {
-                context?.toast(R.string.download_area_too_big, Toast.LENGTH_LONG)
-            } else {
-                if (downloadController.isPriorityDownloadInProgress) {
-                    context?.let {
-                        AlertDialog.Builder(it)
-                            .setMessage(R.string.confirmation_cancel_prev_download_title)
-                            .setPositiveButton(R.string.confirmation_cancel_prev_download_confirmed) { _, _ ->
-                                downloadAreaConfirmed(enclosingBBox)
-                            }
-                            .setNegativeButton(R.string.confirmation_cancel_prev_download_cancel, null)
-                            .show()
-                    }
-                } else {
-                    downloadAreaConfirmed(enclosingBBox)
-                }
-            }
-        }
-    }
+        val downloadArea = listener?.getDownloadArea() ?: return
 
-    private fun downloadAreaConfirmed(bbox: BoundingBox) {
-        var bbox = bbox
-        val areaInSqKm = bbox.area() / 1000000
-        // below a certain threshold, it does not make sense to download, so let's enlarge it
-        if (areaInSqKm < ApplicationConstants.MIN_DOWNLOADABLE_AREA_IN_SQKM) {
-            val cameraPosition = (requireParentFragment() as MainFragment).mapFragment?.cameraPosition
-            if (cameraPosition != null) {
-                val radius = sqrt( 1000000 * ApplicationConstants.MIN_DOWNLOADABLE_AREA_IN_SQKM / PI)
-                bbox = cameraPosition.position.enclosingBoundingBox(radius)
+        if (downloadController.isPriorityDownloadInProgress) {
+            context?.let {
+                AlertDialog.Builder(it)
+                    .setMessage(R.string.confirmation_cancel_prev_download_title)
+                    .setPositiveButton(R.string.confirmation_cancel_prev_download_confirmed) { _, _ ->
+                        downloadController.download(downloadArea, true)
+                    }
+                    .setNegativeButton(R.string.confirmation_cancel_prev_download_cancel, null)
+                    .show()
             }
+        } else {
+            downloadController.download(downloadArea, true)
         }
-        downloadController.download(bbox, true)
     }
 }
